Move all Drag and Animations components to new file
